Add shallow_copy() and use it in ToTree stream sink
Previously, the ToTree sink would add the nodes it receives into a
tree, no matter where they came from. This causes parentship issues
and is not quite nice.

Now, the sink checks whether the nodes are un-owned, and if so, it
uses them. If not, it creates shallow copies and uses these.